+++ src/main/java/com/norab/movie/MovieId.java
@@ -1,4 +1,4 @@
 package com.norab.movie;
 
-public record MovieId(Long movie_id) {
+public record MovieId(Integer movie_id) {
 }
